Code now uses double instead of float

+++ src/torchscript_nn_in_c/c_wrapper.cpp
@@ -13,7 +13,7 @@ void load_model(const char* model_path) {
     NN_Model_instance = new NN_Model(model_path);
 }
 
-float* run_model(float data[INPUT_PIXEL_SIZE][INPUT_PIXEL_SIZE]) {
+double* run_model(double data[INPUT_PIXEL_SIZE][INPUT_PIXEL_SIZE]) {
     return NN_Model_instance->run_model(data);
 }
